Fix: Update services naming convention, and abstracted database seeding/migration.
